colocating loaders with their components
