+++ src/applications/accredited-representative-portal/containers/POARequestDetailsPage.jsx
@@ -1,7 +1,10 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import { Link, useLoaderData } from 'react-router-dom';
+
+import { apiRequest } from '@department-of-veterans-affairs/platform-utilities/api';
 import { formatDateShort } from 'platform/utilities/date/index';
+import mockPOARequestsResponse from '../mocks/mockPOARequestsResponse.json';
 
 const checkAuthorizations = (
   isTreatmentDisclosureAuthorized,
@@ -162,7 +165,26 @@ const POARequestDetailsPage = () => {
     </section>
   );
 };
+
 POARequestDetailsPage.propTypes = {
   usePOARequests: PropTypes.func.isRequired,
 };
+
 export default POARequestDetailsPage;
+
+export async function poaRequestLoader({ params }) {
+  const { id } = params;
+
+  try {
+    const response = await apiRequest(`/power_of_attorney_request/${id}`, {
+      apiVersion: 'accredited_representative_portal/v0',
+    });
+    return response.data;
+  } catch (error) {
+    // Return mock data if API fails (TODO: remove this before pilot and replace with commented throw below)
+    // throwing the error will cause the app to show the error message configured in routes.jsx
+    return mockPOARequestsResponse.data.find(r => r.id === Number(id))
+      ?.attributes;
+    // throw error;
+  }
+}
