+++ src/applications/accredited-representative-portal/containers/POARequestsPage.jsx
@@ -1,8 +1,10 @@
 import React from 'react';
 import { useLoaderData } from 'react-router-dom';
 
+import { apiRequest } from '@department-of-veterans-affairs/platform-utilities/api';
 import DigitalSubmissionAlert from '../components/DigitalSubmissionAlert/DigitalSubmissionAlert';
 import POARequestsCard from '../components/POARequestsCard/POARequestsCard';
+import mockPOARequestsResponse from '../mocks/mockPOARequestsResponse.json';
 
 const POARequestsPage = () => {
   const poaRequests = useLoaderData();
@@ -26,3 +28,17 @@ const POARequestsPage = () => {
 };
 
 export default POARequestsPage;
+
+export async function poaRequestsLoader() {
+  try {
+    const response = await apiRequest('/power_of_attorney_requests', {
+      apiVersion: 'accredited_representative_portal/v0',
+    });
+    return response.data;
+  } catch (error) {
+    // Return mock data if API fails
+    // TODO: Remove mock data before pilot and uncomment throw statement
+    return mockPOARequestsResponse.data;
+    // throw error;
+  }
+}
